feat: move snackbar into design system
