+++ apps/eodh-fe/src/app/app.component.tsx
@@ -1,4 +1,4 @@
-import { AppLoader } from '@ukri/shared/design-system';
+import { AppLoader, SnackbarProvider } from '@ukri/shared/design-system';
 import { OnboardingProvider } from '@ukri/shared/ui/ac-workflow-onboarding';
 import { AuthInterceptor, AuthProvider, KeycloakAdapter } from '@ukri/shared/utils/authorization';
 import { initHttpClient, withQueryClient } from '@ukri/shared/utils/react-query';
@@ -8,7 +8,6 @@ import { I18nextProvider } from 'react-i18next';
 import { getEnvConfig } from '../env.config';
 import i18n from '../i18n';
 import { DefaultLayout } from './layout/default-layout.component';
-import { SnackbarProvider } from './snackbar-provider.component';
 
 const keycloakAdapter = new KeycloakAdapter(getEnvConfig().module.authorization);
