+++ apps/storybook-host/.storybook/preview.tsx
@@ -1,9 +1,9 @@
 import './main.css';
 
 import { QueryClientProvider } from '@tanstack/react-query';
+import { SnackbarProvider } from '@ukri/shared/design-system';
 import { AuthInterceptor, AuthProvider, KeycloakAdapter } from '@ukri/shared/utils/authorization';
 import { initHttpClient, queryClient } from '@ukri/shared/utils/react-query';
-import { SnackbarProvider } from 'notistack';
 import { ComponentType, Suspense, useEffect } from 'react';
 import { I18nextProvider } from 'react-i18next';
