+++ libs/shared/design-system/src/index.ts
@@ -23,3 +23,5 @@ export { Error } from './lib/error/error';
 export { ResultItem, type IResultItemProps } from './lib/result-item/result-item';
 export { Tooltip } from './lib/tooltip/tooltip';
 export { useOutsideClick } from './lib/hooks/use-outside-click';
+export { SnackbarProvider } from './lib/snackbar/snackbar-provider.component';
+export { SnackbarCloseButton } from './lib/snackbar/snackbar-close-button.component';
